requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
bidict==0.23.1
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
deprecation==2.1.0
fsspec==2026.2.0
granian==2.7.2
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
packaging==26.0
platformdirs==4.9.4
postgrest==2.28.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-frontmatter==1.1.0
python-multipart==0.0.22
python-socketio==5.16.1
PyYAML==6.0.3
realtime==2.28.0
redis==7.2.1
resend==2.27.0
reflex==0.8.27
reflex-hosting-cli==0.1.61
requests==2.32.5
rich==14.3.3
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.1.1
wsproto==1.3.2
yarl==1.23.0
zstandard==0.25.0
